start moving files to src/debugger
