+++ src/debugger/map_parser.h
@@ -1,0 +1,3 @@
+#pragma once
+
+char *search_symbol(void *vaddr, int *line);
