--- src/game/map_parser.c
@@ -1,901 +1,0 @@
-#include <ultra64.h>
-#include <macros.h>
-#include <PR/os_internal_error.h>
-#include <stdarg.h>
-#include <string.h>
-#include "segments.h"
-#include "memory.h"
-#include "backtrace.h"
-#include "cop0.h"
-
-extern u32 strstr(char *, char *);
-char *search_symbol(void *vaddr, int *line);
-
-/**
- * @brief Exception codes
- */
-typedef enum {
-    EXCEPTION_CODE_INTERRUPT = 0,
-    EXCEPTION_CODE_TLB_MODIFICATION = 1,
-    EXCEPTION_CODE_TLB_LOAD_I_MISS = 2,
-    EXCEPTION_CODE_TLB_STORE_MISS = 3,
-    EXCEPTION_CODE_LOAD_I_ADDRESS_ERROR = 4,
-    EXCEPTION_CODE_STORE_ADDRESS_ERROR = 5,
-    EXCEPTION_CODE_I_BUS_ERROR = 6,
-    EXCEPTION_CODE_D_BUS_ERROR = 7,
-    EXCEPTION_CODE_SYS_CALL = 8,
-    EXCEPTION_CODE_BREAKPOINT = 9,
-    EXCEPTION_CODE_RESERVED_INSTRUCTION = 10,
-    EXCEPTION_CODE_COPROCESSOR_UNUSABLE = 11,
-    EXCEPTION_CODE_ARITHMETIC_OVERFLOW = 12,
-    EXCEPTION_CODE_TRAP = 13,
-    EXCEPTION_CODE_FLOATING_POINT = 15,
-    EXCEPTION_CODE_WATCH = 23,
-} exception_code_t;
-
-/** @brief Check if addr is a valid PC address */
-static u8 is_valid_address(u32 addr)
-{
-    // TODO: for now we only handle RAM (cached access). This should be extended to handle
-    // TLB-mapped addresses for instance.
-    return addr >= 0x80000400 && addr < 0x80800000 && (addr & 3) == 0;
-}
-
-#define offsetof(type, member)  __builtin_offsetof (type, member)
-
-// The start and end of the exception vector
-extern u8 __osExceptionPreamble[];
-extern u8 send_mesg[];
-#define is_in_exception(addr) ((addr) >= (u32)__osExceptionPreamble && (addr) < (u32)send_mesg)
-
-// code provided by Wiseguy
-static void headless_dma(u32 devAddr, void *dramAddr, u32 size)
-{
-    register u32 stat = IO_READ(PI_STATUS_REG);
-    while (stat & (PI_STATUS_IO_BUSY | PI_STATUS_DMA_BUSY)) {
-        stat = IO_READ(PI_STATUS_REG);
-    }
-    IO_WRITE(PI_DRAM_ADDR_REG, K0_TO_PHYS(dramAddr));
-    IO_WRITE(PI_CART_ADDR_REG, K1_TO_PHYS((u32)osRomBase | devAddr));
-    IO_WRITE(PI_WR_LEN_REG, size - 1);
-}
-static u32 headless_pi_status(void)
-{
-    return IO_READ(PI_STATUS_REG);
-}
-// end of code provided by Wiseguy
-
-/** @brief Address of the SYMT symbol table in the rompak. */
-static u32 SYMT_ROM = 0xFFFFFFFF;
-
-/** @brief Placeholder used in frames where symbols are not available */
-static const char *UNKNOWN_SYMBOL = "???";
-
-extern void dma_read(void *dest, void *srcStart, void *srcEnd);
-
-void map_parser_dma(void *dst, void *src, size_t size) {
-    // char dm[500];
-    // sprintf(dm, "DMA [%08X] <-- [%08X %08X] (%08X)", dst, src, size, SYMT_ROM);
-    // osSyncPrintf(dm);
-    // dma_read(dst, src, src+ALIGN8(size));
-    headless_dma((u32)src, dst, size);
-    while (headless_pi_status() & PI_STATUS_IO_BUSY);
-}
-
-#define STACK_TRAVERSAL_LIMIT 100
-
-/** @brief Function alignment enfored by the compiler (-falign-functions). 
- * 
- * @note This must be kept in sync with n64.mk.
- */
-#define FUNCTION_ALIGNMENT      32
-
-extern u8 _mapDataSegmentRomStart[];
-
-extern u8 _mainSegmentStart[];
-extern u8 _mainSegmentTextEnd[];
-extern u8 _engineSegmentStart[];
-extern u8 _engineSegmentTextEnd[];
-extern u8 _goddardSegmentStart[];
-extern u8 _goddardSegmentTextEnd[];
-
-/** 
- * @brief Symbol table file header
- * 
- * The SYMT file is made of three main tables:
- * 
- * * Address table: this is a sequence of 32-bit integers, each representing an address in the ROM.
- *   The table is sorted in ascending order to allow for binary search. Moreover, the lowest 2 bits
- *   of each address can store additional information: If bit 0 is set to 1, the address is the start
- *   of a function. If bit 1 is set to 1, the address is an inline duplicate. In fact, there might be
- *   multiple symbols at the same address for inlined functions, so we need one entry in this table
- *   for each entry; all of them will have the same address, and all but the last one will have bit
- *   1 set to 1.
- * * Symbol table: this is a sequence of symbol table entries, each representing a symbol. The size
- *   of this table (in number of entries) is exactly the same as the address table. In fact, each
- *   address of the address table can be thought of as an external member of this structure; it's
- *   split externally to allow for efficiency reasons. Each entry stores the function name, 
- *   the source file name and line number, and the binary offset of the symbol within the containing
- *   function.
- * * String table: this table can be thought as a large buffer holding all the strings needed by all
- *   symbol entries (function names and file names). Each symbol entry stores a string as an offset
- *   within the symbol table and a length. This allows to reuse the same string (or prefix thereof)
- *   multiple times. Notice that strings are not null terminated in the string table.
- * 
- * The SYMT file is generated by the n64sym tool during the build process.
- */
-typedef struct {
-    char head[4];           ///< Magic ID "SYMT"
-    u32 version;       ///< Version of the symbol table
-    u32 addrtab_off;   ///< Offset of the address table in the file
-    u32 addrtab_size;  ///< Size of the address table in the file (number of entries)
-    u32 symtab_off;    ///< Offset of the symbol table in the file
-    u32 symtab_size;   ///< Size of the symbol table in the file (number of entries); always equal to addrtab_size.
-    u32 strtab_off;    ///< Offset of the string table in the file
-    u32 strtab_size;   ///< Size of the string table in the file (number of entries)
-} symtable_header_t;
-
-void print_header(symtable_header_t *h) {
-    char tt[1000];
-    char *tp = tt;
-
-    tp += sprintf(tp, "HEAD: %s\n", h->head);
-    tp += sprintf(tp, "VER: %08X\n", h->version);
-    #define PRINT(field) tp += sprintf(tp, #field ": %08X\n", field);
-    PRINT(h->addrtab_off);
-    PRINT(h->addrtab_size);
-    PRINT(h->symtab_off);
-    PRINT(h->symtab_size);
-    PRINT(h->strtab_off);
-    PRINT(h->strtab_size);
-
-    osSyncPrintf(tt);
-}
-
-u8 symHeap[0x4000];
-u32 symIdx = 0;
-void *malloc(size_t size) {
-    return &symHeap[symIdx += ALIGN16(size)];
-}
-
-/** @brief Symbol table entry **/
-typedef struct {
-    u32 func_sidx;     ///< Offset of the function name in the string table
-    u32 file_sidx;     ///< Offset of the file name in the string table
-    u16 func_len;      ///< Length of the function name
-    u16 file_len;      ///< Length of the file name
-    u16 line;          ///< Line number (or 0 if this symbol generically refers to a whole function)
-    u16 func_off;      ///< Offset of the symbol within its function
-} symtable_entry_t;
-
-void print_entry(symtable_entry_t *h) {
-    char tt[1000];
-    char *tp = tt;
-
-    #define PRINT(field) tp += sprintf(tp, #field ": %08X\n", field);
-    PRINT(h->func_sidx);
-    PRINT(h->file_sidx);
-    PRINT(h->func_len);
-    PRINT(h->file_len);
-    PRINT(h->line);
-    PRINT(h->func_off);
-
-    osSyncPrintf(tt);
-}
-
-/** 
- * @brief Entry in the address table.
- * 
- * This is an address in RAM, with the lowest 2 bits used to store additional information.
- * See the ADDRENTRY_* macros to access the various components.
- */
-typedef u32 addrtable_entry_t;
-
-#define ADDRENTRY_ADDR(e)       ((e) & ~3)     ///< Address (without the flags)
-#define ADDRENTRY_IS_FUNC(e)    ((e) &  1)     ///< TRUE if the address is the start of a function
-#define ADDRENTRY_IS_INLINE(e)  ((e) &  2)     ///< TRUE if the address is an inline duplicate
-
-#define MIPS_OP_ADDIU_SP(op)   (((op) & 0xFFFF0000) == 0x27BD0000)   ///< Matches: addiu $sp, $sp, imm
-#define MIPS_OP_DADDIU_SP(op)  (((op) & 0xFFFF0000) == 0x67BD0000)   ///< Matches: daddiu $sp, $sp, imm
-#define MIPS_OP_JR_RA(op)      (((op) & 0xFFFFFFFF) == 0x03E00008)   ///< Matches: jr $ra
-#define MIPS_OP_SD_RA_SP(op)   (((op) & 0xFFFF0000) == 0xAFBF0000)   ///< Matches: sw $ra, imm($sp)
-#define MIPS_OP_SD_FP_SP(op)   (((op) & 0xFFFF0000) == 0xAFBE0000)   ///< Matches: sw $fp, imm($sp)
-#define MIPS_OP_LUI_GP(op)     (((op) & 0xFFFF0000) == 0x3C1C0000)   ///< Matches: lui $gp, imm
-#define MIPS_OP_NOP(op)        ((op) == 0x00000000)                  ///< Matches: nop
-#define MIPS_OP_MOVE_FP_SP(op) ((op) == 0x03A0F025)                  ///< Matches: move $fp, $sp
-
-/** 
- * @brief Open the SYMT symbol table in the rompak.
- * 
- * If not found, return a null header.
- */
-static symtable_header_t symt_open(void *addr) {
-    SYMT_ROM = (u32)_mapDataSegmentRomStart;
-
-    // We don't _need_ to pass by value for this right
-    symtable_header_t ALIGNED8 symt_header;
-    
-    if (SYMT_ROM == 0) {
-        return (symtable_header_t){0};
-    }
-
-    osWritebackDCache(&symt_header, sizeof(symt_header));
-    map_parser_dma(
-        &symt_header,
-        (uintptr_t *)SYMT_ROM,
-        sizeof(symtable_header_t)
-    );
-
-    if (symt_header.head[0] != 'S' || symt_header.head[1] != 'Y' || symt_header.head[2] != 'M' || symt_header.head[3] != 'T') {
-        osSyncPrintf("backtrace: invalid symbol table found at 0x%08lx\n", SYMT_ROM);
-        SYMT_ROM = 0;
-        return (symtable_header_t){0};
-    }
-    if (symt_header.version != 2) {
-        osSyncPrintf("backtrace: unsupported symbol table version %ld -- please update your n64sym tool\n", symt_header.version);
-        SYMT_ROM = 0;
-        return (symtable_header_t){0};
-    }
-
-    return symt_header;
-}
-
-/**
- * @brief Return an entry in the address table by index
- * 
- * @param symt      SYMT file header
- * @param idx       Index of the entry to return
- * @return addrtable_entry_t  Entry of the address table
- */
-static addrtable_entry_t symt_addrtab_entry(symtable_header_t *symt, int idx)
-{
-    return IO_READ(0xB0000000 | (SYMT_ROM + symt->addrtab_off + idx * 4));
-}
-
-/**
- * @brief Search the SYMT address table for the given address.
- * 
- * Run a binary search to find the entry in the table. If there is a single exact match,
- * the entry is returned. If there are multiple entries with the same address, the first
- * entry is returned (this is the case for inlined functions: so some entries following
- * the current one will have the same address). If there is no exact match, the entry
- * with the biggest address just before the given address is returned.
- *
- * @param symt      SYMT file header
- * @param addr      Address to search for
- * @param idx       If not null, will be set to the index of the entry found (or the index just before)
- * @return          The found entry (or the entry just before)
- */
-static addrtable_entry_t symt_addrtab_search(symtable_header_t *symt, u32 addr, int *idx)
-{
-    int min = 0;
-    int max = symt->addrtab_size - 1;
-    while (min < max) {
-        int mid = (min + max) / 2;
-        addrtable_entry_t entry = symt_addrtab_entry(symt, mid);
-        if (addr <= ADDRENTRY_ADDR(entry)) {
-            max = mid;
-        } else {
-            min = mid + 1;
-        }
-    }
-    addrtable_entry_t entry = symt_addrtab_entry(symt, min);
-    if (min > 0 && ADDRENTRY_ADDR(entry) > addr) {
-        entry = symt_addrtab_entry(symt, --min);
-    }
-    if (idx) {
-        *idx = min;
-    }
-    return entry;
-}
-
-
-/**
- * @brief Fetch a string from the string table
- * 
- * @param symt  SYMT file
- * @param sidx  Index of the first character of the string in the string table
- * @param slen  Length of the string
- * @param buf   Destination buffer
- * @param size  Size of the destination buffer
- * @return char*  Fetched string within the destination buffer (might not be at offset 0 for alignment reasons)
- */
-static char* symt_string(symtable_header_t *symt, int sidx, int slen, char *buf, int size)
-{
-    // Align 2-byte phase of the RAM buffer with the ROM address. This is required
-    // for map_parser_dma.
-    int tweak = (sidx ^ (u32)buf) & 1;
-    char *func = buf + tweak; size -= tweak;
-    int nbytes = MIN(slen, size);
-
-    osWritebackDCache(buf, size);
-    map_parser_dma(
-        func, 
-        (uintptr_t *)(SYMT_ROM + symt->strtab_off + sidx),
-        nbytes
-    );
-    func[nbytes] = 0;
-
-    if (tweak) {
-        buf[0] = ' ';
-    }
-    return func;
-}
-
-/**
- * @brief Fetch a symbol table entry from the SYMT file.
- * 
- * @param symt    SYMT file
- * @param entry   Output entry pointer
- * @param idx     Index of the entry to fetch
- */
-static void symt_entry_fetch(symtable_header_t *symt, symtable_entry_t *entry, int idx)
-{
-    osWritebackDCache(entry, sizeof(symtable_entry_t));
-
-    // char dbg[100];
-    // sprintf(dbg,"symt_entry_fetch %d\n", idx);
-    // osSyncPrintf(dbg);
-    map_parser_dma(
-        entry,
-        (uintptr_t *)(SYMT_ROM + symt->symtab_off + idx * sizeof(symtable_entry_t)),
-        sizeof(symtable_entry_t)
-    );
-}
-
-// Fetch the function name of an entry
-static char* symt_entry_func(symtable_header_t *symt, symtable_entry_t *entry, u32 addr, char *buf, int size)
-{
-    if (is_in_exception(addr)) {
-        // Special case exception handlers. This is just to show something slightly
-        // more readable instead of "notcart+0x0" or similar assembly symbols
-        sprintf(buf, "<EXCEPTION HANDLER>");
-        return buf;
-    } else {
-        return symt_string(symt, entry->func_sidx, entry->func_len, buf, size);
-    }
-}
-
-// Fetch the file name of an entry
-static char* symt_entry_file(symtable_header_t *symt, symtable_entry_t *entry, u32 addr, char *buf, int size)
-{
-    return symt_string(symt, entry->file_sidx, entry->file_len, buf, size);
-}
-
-char* __symbolize(void *vaddr, char *buf, int size) {
-    symtable_header_t symt = symt_open(vaddr);
-    if (symt.head[0]) {
-        u32 addr = (u32)vaddr;
-        int idx = 0;
-        addrtable_entry_t a = symt_addrtab_search(&symt, addr, &idx);
-        while (!ADDRENTRY_IS_FUNC(a)) {
-            a = symt_addrtab_entry(&symt, --idx);
-        }
-
-        symtable_entry_t ALIGNED16 entry;
-        // Read the symbol name
-        symt_entry_fetch(&symt, &entry, idx);
-        char *func = symt_entry_func(&symt, &entry, addr, buf, size-12);
-        char lbuf[12];
-        sprintf(lbuf, "+0x%lx", addr - ADDRENTRY_ADDR(a));
-        return strcat(func, lbuf);
-    }
-    sprintf(buf, "%s", UNKNOWN_SYMBOL);
-    return buf;
-}
-
-char *parse_map(u32 addr) {
-    static char map_name[64] ALIGNED16;
-    __symbolize((u32*)addr, map_name, sizeof(map_name));
-
-    return map_name;
-}
-
-char *get_filename(u32 addr) {
-    static char file_name[64] ALIGNED16;
-
-    return file_name;
-}
-
-char *find_function_in_stack(u32 *sp, int *line) {
-    char *ret = NULL;
-    u32 val = *sp;
-
-    #define CALLSITE_OFFSET 8
-    for (int i = 0; i < STACK_TRAVERSAL_LIMIT; i++) {
-        val = *sp;
-        val = *(u32 *)val;
-        *sp += 4;
-
-
-        if ((val >= (u32)_mainSegmentStart) && (val <= (u32)_mainSegmentTextEnd)) {
-            ret = parse_map(val + CALLSITE_OFFSET);
-            break;
-        }
-        else if ((val >= (u32)_engineSegmentStart) && (val <= (u32)_engineSegmentTextEnd)) {
-            ret = parse_map(val + CALLSITE_OFFSET);
-            break;
-        }
-        else if ((val >= (u32)_goddardSegmentStart) && (val <= (u32)_goddardSegmentTextEnd)) {
-            ret = parse_map(val + CALLSITE_OFFSET);
-            break;
-        }
-    }
-
-    if (ret) {
-        search_symbol((void*)(val + CALLSITE_OFFSET), line);
-    }
-
-    return ret;
-}
-
-
-
-/**
- * @brief Analyze a function to find out its stack frame layout and properties (useful for backtracing).
- * 
- * This function implements the core heuristic used by the backtrace engine. It analyzes the actual
- * code of a function in memory instruction by instruction, trying to find out whether the function
- * uses a stack frame or not, whether it uses a frame pointer, and where the return address is stored.
- * 
- * Since we do not have DWARF informations or similar metadata, we can just do educated guesses. A
- * mistake in the heuristic will result probably in a wrong backtrace from this point on.
- * 
- * The heuristic works as follows:
- * 
- *  * Most functions do have a stack frame. In fact, 99.99% of the functions you can find in a call stack
- *    must have a stack frame, because the only functions without a stack frame are leaf functions (functions
- *    that do not call other functions), which in turns can never be part of a stack trace.
- *  * The heuristic walks the function code backwards, looking for the stack frame. Specifically, it looks
- *    for an instruction saving the RA register to the stack (eg: `sd $ra, nn($sp)`), and an instruction
- *    creating the stack frame (eg: `addiu $sp, $sp, -nn`). Once both are found, the heuristic knows how to
- *    fill in `.stack_size` and `.ra_offset` fields of the function description structure, and it can stop.
- *  * Some functions also modify $fp (the frame pointer register): sometimes, they just use it as one additional
- *    free register, and other times they really use it as frame pointer. If the heuristic finds the
- *    instruction `move $fp, $sp`, it knows that the function uses $fp as frame pointer, and will mark
- *    the function as BT_FUNCTION_FRAMEPOINTER. In any case, the field `.fp_offset` will be filled in
- *    with the offset in the stack where $fp is stored, so that the backtrace engine can track the
- *    current value of the register in any case.
- *  * The 0.01% of the functions that do not have a stack frame but appear in the call stack are leaf
- *    functions interrupted by exceptions. Leaf functions pose two important problems: first, $ra is
- *    not saved into the stack so there is no way to know where to go back. Second, there is no clear
- *    indication where the function begins (as we normally stops analysis when we see the stack frame
- *    creation). So in this case the heuristic would fail. We rely thus on two hints coming from the caller:
- *    * First, we expect the caller to set from_exception=TRUE, so that we know that we might potentially
- *      deal with a leaf function. 
- *    * Second, the caller should provide the function start address, so that we stop the analysis when
- *      we reach it, and mark the function as BT_LEAF.
- *    * If the function start address is not provided (because e.g. the symbol table was not found and
- *      thus we have no information about function starts), the last ditch heuristic is to look for
- *      the nops that are normally used to align the function start to the FUNCTION_ALIGNMENT boundary.
- *      Obviously this is a very fragile heuristic (it will fail if the function required no nops to be
- *      properly aligned), but it is the best we can do. Worst case, in this specific case of a leaf
- *      function interrupted by the exception, the stack trace will be wrong from this point on.
- * 
- * @param func                        Output function description structure
- * @param ptr                         Pointer to the function code at the point where the backtrace starts.
- *                                    This is normally the point where a JAL opcode is found, as we are walking
- *                                    up the call stack.
- * @param func_start                  Start of the function being analyzed. This is optional: the heuristic can work
- *                                    without this hint, but it is useful in certain situations (eg: to better
- *                                    walk up after an exception).
- * @param from_exception              If TRUE, this function was interrupted by an exception. This is a hint that
- *                                    the function *might* even be a leaf function without a stack frame, and that
- *                                    we must use special heuristics for it.
- * 
- * @return TRUE if the backtrace can continue, FALSE if must be aborted (eg: we are within invalid memory)
- */
-u8 __bt_analyze_func(bt_func_t *func, u32 *ptr, u32 func_start, u8 from_exception)
-{
-    *func = (bt_func_t){
-        .type = is_in_exception(ptr) ? BT_EXCEPTION : BT_FUNCTION,
-        .stack_size = 0, .ra_offset = 0, .fp_offset = 0
-    };
-
-    u32 addr = (u32)ptr;
-    while (1) {
-        // Validate that we can dereference the virtual address without raising an exception
-        // TODO: enhance this check with more valid ranges.
-        if (!is_valid_address(addr)) {
-            // This address is invalid, probably something is corrupted. Avoid looking further.
-            osSyncPrintf("backtrace: interrupted because of invalid return address 0x%08lx\n", addr);
-            return FALSE;
-        }
-        u32 op = *(u32*)addr;
-        if (MIPS_OP_ADDIU_SP(op) || MIPS_OP_DADDIU_SP(op)) {
-            // Extract the stack size only from the start of the function, where the
-            // stack is allocated (negative value). This is important because the RA
-            // could point to a leaf basis block at the end of the function (like in the
-            // assert case), and if we picked the positive ADDIU SP at the end of the
-            // proper function body, we might miss a fp_offset.
-            if (op & 0x8000)
-                func->stack_size = -(s16)(op & 0xFFFF);
-        } else if (MIPS_OP_SD_RA_SP(op)) {
-            osSyncPrintf("sw ra,sp");
-            func->ra_offset = (s16)(op & 0xFFFF) + 4; // +4 = load low 32 bit of RA
-            // If we found a stack size, it might be a red herring (an alloca); we need one
-            // happening "just before" sd ra,xx(sp)
-            func->stack_size = 0;
-        } else if (MIPS_OP_SD_FP_SP(op)) {
-            func->fp_offset = (s16)(op & 0xFFFF) + 4; // +4 = load low 32 bit of FP
-        } else if (MIPS_OP_LUI_GP(op)) {
-            // Loading gp is commonly done in _start, so it's useless to go back more
-            return FALSE;
-        } else if (MIPS_OP_MOVE_FP_SP(op)) {
-            // This function uses the frame pointer. Uses that as base of the stack.
-            // Even with -fomit-frame-pointer (default on our toolchain), the compiler
-            // still emits a framepointer for functions using a variable stack size
-            // (eg: using alloca() or VLAs).
-            func->type = BT_FUNCTION_FRAMEPOINTER;
-        } 
-        // We found the stack frame size and the offset of the return address in the stack frame
-        // We can stop looking and process the frame
-        if (func->stack_size != 0 && func->ra_offset != 0)
-            break;
-        if (from_exception) {
-            // The function we are analyzing was interrupted by an exception, so it might
-            // potentially be a leaf function (no stack frame). We need to make sure to stop
-            // at the beginning of the function and mark it as leaf function. Use
-            // func_start if specified, or try to guess using the nops used to align the function
-            // (crossing fingers that they're there).
-            if (addr == func_start) {
-                // The frame that was interrupted by an interrupt handler is a special case: the
-                // function could be a leaf function with no stack. If we were able to identify
-                // the function start (via the symbol table) and we reach it, it means that
-                // we are in a real leaf function.
-                func->type = BT_LEAF;
-                break;
-            } else if (!func_start && MIPS_OP_NOP(op) && (addr + 4) % FUNCTION_ALIGNMENT == 0) {
-                // If we are in the frame interrupted by an interrupt handler, and we does not know
-                // the start of the function (eg: no symbol table), then try to stop by looking for
-                // a NOP that pads between functions. Obviously the NOP we find can be either a FALSE
-                // positive or a FALSE negative, but we can't do any better without symbols.
-                func->type = BT_LEAF;
-                break;
-            }
-        }
-        addr -= 4;
-    }
-    return TRUE;
-}
-
-static void backtrace_foreach(void (*cb)(void *arg, void *ptr), void *arg, u32 *ra, u32 *sp, u32 *fp, u32 *exception_ra)
-{
-    /*
-     * This function is called in very risky contexts, for instance as part of an exception
-     * handler or during an assertion. We try to always provide as much information as
-     * possible in these cases, with graceful degradation if something more elaborate cannot
-     * be extracted. Thus, this function:
-     * 
-     *  * Must not use malloc(). The heap might be corrupted or empty.
-     *  * Must not use assert(), because that might trigger recursive assertions.
-     *  * Must avoid raising exceptions. Specifically, it must avoid risky memory accesses
-     *    to wrong addresses.
-     */
-
-    // #if BACKTRACE_DEBUG
-    osSyncPrintf("backtrace: start\n"); 
-    // #endif
-
-    u32 func_start = 0;            // Start of the current function (when known)
-
-    // Start calling the callback for the backtrace entry point
-    cb(arg, ra);
-
-    while (1) {
-        // Analyze the function pointed by ra, passing information about the previous exception frame if any.
-        // If the analysis fail (for invalid memory accesses), stop right away.
-        bt_func_t func; 
-        if (!__bt_analyze_func(&func, ra, func_start, (u8)exception_ra))
-            return;
-
-        #if 1
-        char debug[200];
-        sprintf(debug, "backtrace: %s, ra=%p, sp=%p, fp=%p ra_offset=%d, fp_offset=%d, stack_size=%d\n", 
-            func.type == BT_FUNCTION ? "BT_FUNCTION" : (func.type == BT_EXCEPTION ? "BT_EXCEPTION" : (func.type == BT_FUNCTION_FRAMEPOINTER ? "BT_FRAMEPOINTER" : "BT_LEAF")),
-            ra, sp, fp, func.ra_offset, func.fp_offset, func.stack_size);
-        osSyncPrintf(debug);
-        #endif
-
-        switch (func.type) {
-            case BT_FUNCTION_FRAMEPOINTER:
-                if (!func.fp_offset) {
-                    osSyncPrintf("backtrace: framepointer used but not saved onto stack at %p\n", ra);
-                } else {
-                    // Use the frame pointer to refer to the current frame.
-                    sp = fp;
-                    if (!is_valid_address((u32)sp)) {
-                        osSyncPrintf("backtrace: interrupted because of invalid frame pointer 0x%08lx\n", (u32)sp);
-                        return;
-                    }
-                }
-                // FALLTHROUGH!
-            case BT_FUNCTION:
-                if (func.fp_offset)
-                    fp = *(u32**)((u32)sp + func.fp_offset);
-                ra = *(u32**)((u32)sp + func.ra_offset) - 2;
-                sp = (u32*)((u32)sp + func.stack_size);
-                exception_ra = NULL;
-                func_start = 0;
-                break;
-            case BT_EXCEPTION: {
-                // Exception frame. We must return back to EPC, but let's keep the
-                // RA value. If the interrupted function is a leaf function, we
-                // will need it to further walk back.
-                // Notice that FP is a callee-saved register so we don't need to
-                // recover it from the exception frame (also, it isn't saved there
-                // during interrupts).
-                exception_ra = *(u32**)((u32)sp + func.ra_offset);
-
-                // Read EPC from exception frame and adjust it with CAUSE BD bit
-                ra = *(u32**)((u32)sp + offsetof(__OSThreadContext, pc) + 32);
-                u32 cause = *(u32*)((u32)sp + offsetof(__OSThreadContext, cause) + 32);
-                if (cause & C0_CAUSE_BD) ra++;
-
-                sp = (u32*)((u32)sp + func.stack_size);
-
-                // Special case: if the exception is due to an invalid EPC
-                // (eg: a null function pointer call), we can rely on RA to get
-                // back to the caller. This assumes that we got there via a function call
-                // rather than a raw jump, but that's a reasonable assumption. It's anyway
-                // the best we can do.
-                if ((C0_GET_CAUSE_EXC_CODE(cause) == EXCEPTION_CODE_TLB_LOAD_I_MISS ||
-                    C0_GET_CAUSE_EXC_CODE(cause) == EXCEPTION_CODE_LOAD_I_ADDRESS_ERROR) && 
-                    !is_valid_address((u32)ra)) {
-                    
-                    // Store the invalid address in the backtrace, so that it will appear in dumps.
-                    // This makes it easier for the user to understand the reason for the exception.
-                    cb(arg, ra);
-                    
-                    ra = exception_ra - 2;
-
-                    // The function that jumped into an invalid PC was not interrupted by the exception: it
-                    // is a regular function
-                    // call now.
-                    exception_ra = NULL;
-                    break;
-                }
-
-                // The next frame might be a leaf function, for which we will not be able
-                // to find a stack frame. It is useful to try finding the function start.
-                // Try to open the symbol table: if we find it, we can search for the start
-                // address of the function.
-                symtable_header_t symt = symt_open(ra);
-                if (symt.head[0]) {
-                    int idx;
-                    addrtable_entry_t entry = symt_addrtab_search(&symt, (u32)ra, &idx);
-                    while (!ADDRENTRY_IS_FUNC(entry))
-                        entry = symt_addrtab_entry(&symt, --idx);
-                    func_start = ADDRENTRY_ADDR(entry);
-                }
-            }   break;
-            case BT_LEAF:
-                ra = exception_ra - 2;
-                // A leaf function has no stack. On the other hand, an exception happening at the
-                // beginning of a standard function (before RA is saved), does have a stack but
-                // will be marked as a leaf function. In this case, we mus update the stack pointer.
-                sp = (u32*)((u32)sp + func.stack_size);
-                exception_ra = NULL;
-                func_start = 0;
-                break;
-        }
-
-        // Call the callback with this stack frame
-        cb(arg, ra);
-    }
-}
-
-void _flush_cache() {
-    // no-op
-}
-
-int backtrace(void **buffer, int size) {
-    int i = -1; // Skip backtrace() itself
-    void cb(void *arg, void *ptr) {
-        if (i >= size) return;
-        if (i >= 0)
-            buffer[i] = ptr;
-        i++;
-    }
-
-    // Current value of SP/RA/FP registers.
-    u32 *sp, *fp;
-    asm volatile (
-        "move %0, $sp\n"
-        "move %1, $fp\n"
-        : "=r"(sp), "=r"(fp)
-    );
-
-    // Start from the backtrace function itself. Put the start pointer
-    // somewhere after the initial prolog, so that we parse the prolog
-    // itself to find sp/fp/ra offsets.
-    // Since we don't come from an exception, exception_ra must be NULL.
-    u32 *pc = (u32*)backtrace + 24;
-
-    backtrace_foreach(cb, NULL, pc, sp, fp, NULL);
-    return i;
-}
-
-int __backtrace_from(void **buffer, int size, u32 *pc, u32 *sp, u32 *fp, u32 *exception_ra)
-{
-    int i = 0;
-    void cb(void *arg, void *ptr) {
-        if (i >= size) return;
-        if (i >= 0)
-            buffer[i] = ptr;
-        i++;
-    }
-
-    backtrace_foreach(cb, NULL, pc, sp, fp, exception_ra);
-    return i;
-}
-
-
-static void format_entry(void (*cb)(void *, backtrace_frame_t *), void *cb_arg, 
-    symtable_header_t *symt, int idx, u32 addr, u32 offset, u8 is_func, u8 is_inline)
-{       
-    symtable_entry_t ALIGNED8 entry;
-    symt_entry_fetch(symt, &entry, idx);
-
-    char ALIGNED8 file_buf[entry.file_len + 2];
-    char ALIGNED8 func_buf[MAX(entry.func_len + 2, 32)];
-
-    cb(cb_arg, &(backtrace_frame_t){
-        .addr = addr,
-        .func_offset = offset ? offset : entry.func_off,
-        .func = symt_entry_func(symt, &entry, addr, func_buf, sizeof(func_buf)),
-        .source_file = symt_entry_file(symt, &entry, addr, file_buf, sizeof(file_buf)),
-        .source_line = is_func ? 0 : entry.line,
-        .is_inline = is_inline,
-    });
-}
-
-u8 backtrace_symbols_cb(void **buffer, int size, u32 flags,
-    void (*cb)(void *, backtrace_frame_t *), void *cb_arg)
-{
-    for (int i=0; i<size; i++) {
-        u32 needle = (u32)buffer[i];
-        // Open the symbol table. If not found, we will still invoke the
-        // callback but using unsymbolized addresses.
-        symtable_header_t symt_header = symt_open(buffer[i]);
-        u8 has_symt = symt_header.head[0];
-        if (!is_valid_address(needle)) {
-            // If the address is before the first symbol, we call it a NULL pointer, as that is the most likely case
-            cb(cb_arg, &(backtrace_frame_t){
-                .addr = needle,
-                .func_offset = needle,
-                .func = needle < 128 ? "<NULL POINTER>" : "<INVALID ADDRESS>",
-                .source_file = UNKNOWN_SYMBOL, .source_line = 0, .is_inline = FALSE
-            });
-            continue;
-        }
-        if (!has_symt) {
-            // No symbol table. Call the callback with a dummy entry which just contains the address
-            u8 exc = is_in_exception(needle);
-            cb(cb_arg, &(backtrace_frame_t){
-                .addr = needle,
-                .func = exc ? "<EXCEPTION HANDLER>" : UNKNOWN_SYMBOL, .func_offset = 0,
-                .source_file = UNKNOWN_SYMBOL, .source_line = 0, .is_inline = FALSE
-            });
-            continue;
-        }
-        int idx; addrtable_entry_t a;
-        a = symt_addrtab_search(&symt_header, needle, &idx);
-
-        if (ADDRENTRY_ADDR(a) == needle) {
-            // Found an entry at this address. Go through all inlines for this address.
-            while (1) {
-                format_entry(cb, cb_arg, &symt_header, idx, needle, 0, FALSE, ADDRENTRY_IS_INLINE(a));
-                if (!ADDRENTRY_IS_INLINE(a)) break;
-                a = symt_addrtab_entry(&symt_header, ++idx);
-            }
-        } else {
-            // Search the containing function
-            while (!ADDRENTRY_IS_FUNC(a))
-                a = symt_addrtab_entry(&symt_header, --idx);
-            format_entry(cb, cb_arg, &symt_header, idx, needle, needle - ADDRENTRY_ADDR(a), TRUE, FALSE);
-        }
-    }
-    return TRUE;
-}
-
-char** backtrace_symbols(void **buffer, int size)
-{
-    const int MAX_FILE_LEN = 120;
-    const int MAX_FUNC_LEN = 120;
-    const int MAX_SYM_LEN = MAX_FILE_LEN + MAX_FUNC_LEN + 24;
-    char **syms = malloc(2 * size * (sizeof(char*) + MAX_SYM_LEN));
-    char *out = (char*)syms + size*sizeof(char*);
-    int level = 0;
-
-    void cb(void *arg, backtrace_frame_t *frame) {
-        int n = sprintf(out,
-            "%s+0x%lx (%s:%d) [0x%08lx]", frame->func, frame->func_offset, frame->source_file, frame->source_line, frame->addr);
-        if (frame->is_inline)
-            out[-1] = '\n';
-        else
-            syms[level++] = out;
-        out += n + 1;
-    }
-
-    backtrace_symbols_cb(buffer, size, 0, cb, NULL);
-    return syms;
-}
-
-void crash_screen_print(s32 x, s32 y, const char *fmt, ...);
-
-void backtrace_frame_print_compact(backtrace_frame_t *frame, char *out, int width) {
-    const char *source_file = frame->source_file;
-    int len = strlen(frame->func) + strlen(source_file);
-    u8 ellipsed = FALSE;
-    if (len > width && source_file) {
-        source_file += len - (width - 8);
-        ellipsed = TRUE;
-    }
-    if (frame->func != UNKNOWN_SYMBOL) out += sprintf(out, "%s ", frame->func);
-    if (source_file != UNKNOWN_SYMBOL) out += sprintf(out, "(%s%s:%d)", ellipsed ? "..." : "", source_file, frame->source_line);
-    if (frame->func == UNKNOWN_SYMBOL || source_file == UNKNOWN_SYMBOL)
-        out += sprintf(out, "[0x%08lx]", frame->addr);
-    // out += sprintf(out, "\n");
-}
-
-char *search_symbol(void *vaddr, int *line) {
-    static char filebuf[100];
-
-    symtable_header_t symt = symt_open(vaddr);
-    if (symt.head[0]) {
-        u32 addr = (u32)vaddr;
-        int idx = 0;
-        addrtable_entry_t a = symt_addrtab_search(&symt, addr, &idx);
-        // while (!ADDRENTRY_IS_FUNC(a)) {
-        //     a = symt_addrtab_entry(&symt, --idx);
-        // }
-
-        symtable_entry_t ALIGNED16 entry;
-        // Read the symbol name
-
-        filebuf[0] = 0;
-        symt_entry_fetch(&symt, &entry, idx);
-        *line = entry.line;
-        char *file = symt_entry_file(&symt, &entry, addr, filebuf, 100);
-        // char *ret = file;
-        // file += sprintf(file, ":%d", entry.line);
-        return file;
-    }
-    return UNKNOWN_SYMBOL;
-}
-
-
-void inspector_print_backtrace(void *bt, int n, int bt_skip) {
-    crash_screen_print(10, 45, "Backtrace:");
-    char func[128];
-    u8 skip = TRUE;
-    static int yBase = 55;
-    void cb(void *arg, backtrace_frame_t *frame) {
-        if (skip) {
-            if (strstr(frame->func, "<EXCEPTION HANDLER>"))
-                skip = FALSE;
-            return;
-        }
-        if (bt_skip > 0) {
-            bt_skip--;
-            return;
-        }
-        // crash_screen_print(10, "    ");
-        sprintf(func, "%s", frame->func);
-        frame->func = func;
-        char stdout[400];
-        backtrace_frame_print_compact(frame, stdout, 40);
-        crash_screen_print(10, yBase, stdout);
-        yBase += 10;
-    }
-    backtrace_symbols_cb(bt, n, 0, cb, NULL);
-    if (skip) {
-        // we didn't find the exception handler for some reason (eg: missing symbols)
-        // so just print the whole thing
-        skip = FALSE;
-        backtrace_symbols_cb(bt, n, 0, cb, NULL);
-    }
-}
-
-
